user login

+++ NodeJS&Express/routes/articles.js
@@ -3,9 +3,13 @@ const router = express.Router();
 
 //Bring in Article Models
 let Article = require('../models/article');
+//Bring in User Models
+let User = require('../models/user');
+
+
 
 //Add Route
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticated, async (req, res) => {
     res.render('add_article', {
         title: 'Add Article'
     });
@@ -14,7 +18,7 @@ router.get('/add', (req, res) => {
 // Add article- POST Route
 router.post('/add', function (req, res) {
     req.checkBody('title', 'Title is required').notEmpty();
-    req.checkBody('author', 'Author is required').notEmpty();
+    //req.checkBody('author', 'Author is required').notEmpty();
     req.checkBody('body', 'Body is required').notEmpty();
     //Get errors
     const errors = req.validationErrors(req);
@@ -26,7 +30,7 @@ router.post('/add', function (req, res) {
     } else {
         let article = new Article();
         article.title = req.body.title;
-        article.author = req.body.author;
+        article.author = req.user._id;
         article.body = req.body.body;
 
         article.save((err) => {
@@ -43,8 +47,12 @@ router.post('/add', function (req, res) {
 });
 
 //Load edit Form
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, async (req, res) => {
     Article.findById(req.params.id, (err, article) => {
+        if (article.author != req.user._id) {
+            req.flash('danger', 'Not Authorized');
+            res.redirect('/');
+        }
         res.render('edit_article', {
             title: 'Edit Article',
             article: article
@@ -75,22 +83,43 @@ router.post('/edit/:id', (req, res) => {
 
 //Delete
 router.delete('/:id', (req, res) => {
+    if (!req.user._id) {
+        res.status(500).send();
+    }
     let query = { _id: req.params.id }
-    Article.deleteOne(query, err => {
-        if (err) {
-            console.log(err);
+    Article.findById(req.params.id, function (err, article) {
+        if (article.author != req.user._id) {
+            res.status(500).send();
+        } else {
+            Article.deleteOne(query, err => {
+                if (err) {
+                    console.log(err);
+                }
+                res.send('Success');
+            });
         }
-        res.send('Success');
     });
 });
 
 //Get Single Article
 router.get('/:id', (req, res) => {
     Article.findById(req.params.id, (err, article) => {
-        res.render('article', {
-            article: article
+        User.findById(article.author, function (err, user) {
+            res.render('article', {
+                article: article,
+                author: user.name
+            });
         });
     });
+});
+
+// Access Control
+function ensureAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    } else {
+        req.flash('danger', 'Please login');
+        res.redirect('/users/login');
+    }
 }
-);
 module.exports = router;
